Started implementing individual non-merged downloads

+++ main.py
@@ -26,46 +26,51 @@ def main():
     if 'playlist' not in link:
         print('Only enter Youtube playlists, not regular videos or non-Youtube links! Try again!')
     else:
-        p = Playlist(link)
-        playlistTitle = str(p.title).replace(' ', '').replace('\'', '').replace('/', '') + '.mp3'
-        outputPath = 'DownloadedMusic'
-        os.makedirs(outputPath, exist_ok=True)
+        question = input('Do you want to download or download and merge? [d/DM]')
+        if ('d' or 'dm') not in question:
+            print('Response not proper. Please try again!')             
+        else:
+            p = Playlist(link)
+            playlistTitle = str(p.title).replace(' ', '').replace('\'', '').replace('/', '') + '.mp3'
+            outputPath = 'DownloadedMusic'
+            os.makedirs(outputPath, exist_ok=True)
+                    
+            print('\n', playlistTitle, '\n')
+            if question.lower() == 'd':
                 
-        print('\n', playlistTitle, '\n')
-        
-        count = 1
-        for video in p.videos:
-            masterAudioFilePath = os.path.join(outputPath, playlistTitle)
-            print(count, '/', len(p), ': Downloading', video.title)
-            
-            try:
-                audioFile = video.streams.filter(only_audio=True).first()
+                count = 1
+                for video in p.videos:
+                    masterAudioFilePath = os.path.join(outputPath, playlistTitle)
+                    print(count, '/', len(p), ': Downloading', video.title)
+                    
+                    try:
+                        audioFile = video.streams.filter(only_audio=True).first()
+                        
+                        # Creates the master file 
+                        if count == 1:
+                            downloadAudioFile(audioFile, outputPath, playlistTitle)
+                            print("Master audio file created successfully.\n")
+                            
+                        # Adds to the master file
+                        else:
+                            addonFileName = 'addon.mp3'
+                            addonFilePath = os.path.join(outputPath, addonFileName)
+                            downloadAudioFile(audioFile, outputPath, addonFileName)
+                            
+                            mergedFile = mergeAudioFiles(masterAudioFilePath, addonFilePath, outputPath)   
+                                        
+                            # Replaces the outdated master file with the new master file, and deletes the no-longer-needed addon file
+                            os.replace(mergedFile, masterAudioFilePath)
+                            os.remove(addonFilePath)
+                            
+                            
+                    except pytube.exceptions.AgeRestrictedError as e:
+                        print('Skipping', video.title, 'due to error: \"', e, '\" \n')
+                    
+                    count += 1
+                    
                 
-                # Creates the master file 
-                if count == 1:
-                    downloadAudioFile(audioFile, outputPath, playlistTitle)
-                    print("Master audio file created successfully.\n")
-                    
-                # Adds to the master file
-                else:
-                    addonFileName = 'addon.mp3'
-                    addonFilePath = os.path.join(outputPath, addonFileName)
-                    downloadAudioFile(audioFile, outputPath, addonFileName)
-                    
-                    mergedFile = mergeAudioFiles(masterAudioFilePath, addonFilePath, outputPath)   
-                                
-                    # Replaces the outdated master file with the new master file, and deletes the no-longer-needed addon file
-                    os.replace(mergedFile, masterAudioFilePath)
-                    os.remove(addonFilePath)
-                    
-                    
-            except pytube.exceptions.AgeRestrictedError as e:
-                print('Skipping', video.title, 'due to error: \"', e, '\" \n')
-            
-            count += 1
-                
-            
-        print('All files successfully downloaded and merged!\n')
+            print('All files successfully downloaded and merged!\n')
 
 if __name__ == '__main__': 
     main()
